requirements.txt
torch
numpy
matplotlib
notebook
PyYAML
dill
gymnasium[box2d]
pygame